+++ 250226_UMLChess/250226_UMLChess.vcxproj
@@ -128,10 +128,12 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="250226_UMLChess.cpp" />
+    <ClCompile Include="Board.cpp" />
     <ClCompile Include="ChessSaveLoadManager.cpp" />
     <ClCompile Include="Unit.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Board.h" />
     <ClInclude Include="ChessData.h" />
     <ClInclude Include="ChessSaveLoadManager.h" />
     <ClInclude Include="Unit.h" />
